+++ src/TabularDataReader.php
@@ -22,6 +22,8 @@ use IteratorAggregate;
  *
  * @method Iterator fetchColumnByName(string $name)  returns a column from its name
  * @method Iterator fetchColumnByOffset(int $offset) returns a column from its offset
+ * @method array    fetchFirst(): returns the first record from the tabular data.
+ * @method array fetchNth(int $nth_record) returns the nth record from the tabular data.
  */
 interface TabularDataReader extends Countable, IteratorAggregate
 {
@@ -76,6 +78,10 @@ interface TabularDataReader extends Countable, IteratorAggregate
     public function getRecords(array $header = []): Iterator;
 
     /**
+     * DEPRECATION WARNING! This class will be removed in the next major point release.
+     *
+     * @deprecated since version 9.9.0
+     *
      * Returns the nth record from the tabular data.
      *
      * By default, if no index is provided the first record of the tabular data is returned